+++ src/Agents.jl
@@ -29,6 +29,7 @@ include("simulations/sample.jl")
 @deprecate space_neighbors nearby_agents
 @deprecate node_neighbors nearby_positions
 @deprecate get_node_contents agents_in_pos
+@deprecate pick_empty random_empty
 
 # Predefined models
 include("models/Models.jl")