+++ CHANGELOG.md
@@ -18,6 +18,7 @@ therefore are not "truly breaking".
   - `space_neighbors` -> `nearby_agents`
   - `node_neighbors` -> `nearby_positions`
   - `get_node_contents` -> `agents_in_pos`
+  - `pick_empty` -> `random_empty`
 
 # v3.7
 - Add the ability to decide whether the agent step or the model step should be performed first using the `agents_first` argument.